# backend/api/openapi.yaml
openapi: "3.1.0"
info:
    title: "Roshnii API"
    version: "0.1.0" # Updated version
    description: "API for Roshnii - Photo storage and management service."
servers:
    - url: "http://localhost:8080/api" # Base path for API endpoints
      description: "Development server"
paths:
...
    /albums:
        post:
            summary: "Create Album"
            description: "Creates a new album for the authenticated user."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: "The album name."
                                description:
                                    type: string
                                    description: "Optional album description."
                            required:
                                - name
            responses:
                "201":
                    description: "Album created successfully"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Album"
                "400":
                    description: "Bad Request"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        get:
            summary: "List Albums"
            description: "Lists all albums for the authenticated user."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: "List of albums"
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Album"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /albums/{id}:
        get:
            summary: "Get Album"
            description: "Get a specific album by ID."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                  description: "Album ID"
            responses:
                "200":
                    description: "Album details"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Album"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: "Album not found"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        put:
            summary: "Update Album"
            description: "Update an existing album."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                  description: "Album ID"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: "The album name."
                                description:
                                    type: string
                                    description: "Optional album description."
                            required:
                                - name
            responses:
                "200":
                    description: "Album updated successfully"
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                "400":
                    description: "Bad Request"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: "Album not found"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        delete:
            summary: "Delete Album"
            description: "Delete an album and all its image associations."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                  description: "Album ID"
            responses:
                "200":
                    description: "Album deleted successfully"
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: "Album not found"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /albums/{id}/images:
        post:
            summary: "Add Image to Album"
            description: "Add an image to an album."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                  description: "Album ID"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                image_id:
                                    type: string
                                    format: uuid
                                    description: "ID of the image to add."
                            required:
                                - image_id
            responses:
                "200":
                    description: "Image added successfully"
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                "400":
                    description: "Bad Request"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: "Album or image not found"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        get:
            summary: "List Images in Album"
            description: "List all images in an album."
            tags: ["Albums"]
            security:
                - cookieAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                  description: "Album ID"
            responses:
                "200":
                    description: "List of images in album"
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/ImageMetadata"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: "Album not found"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
components:
    schemas:
        ImageMetadata:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                    description: "Unique identifier for the image."
                user_id:
                    type: integer
                    format: int64
                    description: "ID of the user who uploaded the image."
                filename:
                    type: string
                    description: "Original filename of the uploaded image."
                content_type:
                    type: string
                    description: "MIME type of the image (e.g., 'image/jpeg')."
                size:
                    type: integer
                    format: int64
                    description: "Size of the image in bytes."
                width:
                    type: integer
                    description: "Width of the image in pixels (if available)."
                height:
                    type: integer
                    description: "Height of the image in pixels (if available)."
                created_at:
                    type: string
                    format: date-time
                    description: "Timestamp when the image metadata was created."
                updated_at:
                    type: string
                    format: date-time
                    description: "Timestamp when the image metadata was last updated."
            required:
                - id
                - user_id
                - filename
                - content_type
                - size
                - created_at
                - updated_at
...
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
                    description: "A description of the error."
            required:
                - error
        Album:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                    description: "Unique identifier for the album."
                user_id:
                    type: integer
                    format: int64
                    description: "ID of the user who created the album."
                name:
                    type: string
                    description: "Name of the album."
                description:
                    type: string
                    description: "Description of the album (optional)."
                created_at:
                    type: string
                    format: date-time
                    description: "Timestamp when the album was created."
                updated_at:
                    type: string
                    format: date-time
                    description: "Timestamp when the album was last updated."
            required:
                - id
                - user_id
                - name
                - created_at
                - updated_at
    securitySchemes:
        cookieAuth: # Define cookie-based authentication
            type: apiKey
            in: cookie
            name: auth_token # Name of the cookie